Moved remapping of domains of categorical vectors to VecUtils. Removed deep cloning of vecs once remapping occurs from AstReplaceAll and AstReplaceFirst classes.
